requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gsheets==0.6.1
httplib2==0.22.0
idna==3.10
language_tool_python==2.8.1
oauth2client==4.1.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.3.0
six==1.16.0
tqdm==4.66.6
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
